replaced modal component for invite member
